backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==2.0.23
psycopg==3.1.18
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==0.19.0
opencv-python==4.8.1.78
numpy==1.26.2
pillow==10.1.0
requests==2.31.0
